+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Release: dd.mm.yyyy
 #### App
 
 - Change: Function `launch(...)` now supports a focus option to focus the app automatically on launch ([#211](https://github.com/kasper/phoenix/issues/211), [#212](https://github.com/kasper/phoenix/pull/212)).
+- New: Add modifier flags to mouse events ([#216](https://github.com/kasper/phoenix/issues/216)).
 
 2.6.1
 -----